DRY up the two tables and combine into one
